LayoutTests/fast/animation/request-animation-frame-throttle-subframe-display-none-expected.txt

Tests that requestAnimationFrame is not throttled in subframes that are display:none

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testFrame.contentWindow.internals.requestAnimationFrameThrottlingReasons() is "[Unthrottled]"
PASS successfullyParsed is true

TEST COMPLETE
